+++ src/main/java/com/depot/ims/repositories/ShipmentRepository.java
@@ -9,13 +9,13 @@ import java.util.List;
 public interface ShipmentRepository extends JpaRepository<Shipment, Long> {
 
     @Query("select s from Shipment s where s.shipmentId = ?1")
-    Shipment findByShipmentId(Integer shipmentId);
+    Shipment findByShipmentId(Long shipmentId);
 
     @Query("select s from Shipment s where s.source = ?1")
-    List<Shipment> findByShipmentSource(Integer source);
+    List<Shipment> findByShipmentSource(Long source);
 
     @Query("select s from Shipment s where s.destination = ?1")
-    List<Shipment> findByShipmentDestination(Integer destination);
+    List<Shipment> findByShipmentDestination(Long destination);
 
     @Query("select s from Shipment s where s.currentLocation = ?1")
     List<Shipment> findByCurrentLocation(String currentLocation);
